refactor(author): extracts article grid component
